requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
flake8
gunicorn
python-dotenv
requests